Move key handling modules to common dir
